# api/openapi.yaml
openapi: 3.1.0
info:
  title: ochakai REST API
  description: >
    ochakai is a context provider for data agents: metric definitions,
    verified golden queries, interpretation knowledge, glossary terms, and
    table catalog entries in one knowledge base. The REST API mirrors the
    MCP tools so users can build their own web UIs. ochakai executes no SQL
    and uses no LLM.
  version: 0.1.0
  license:
    name: MIT
servers:
  - url: /
# Authentication is Cloud Run IAM (Google identity), handled by the
# infrastructure before requests reach ochakai — there is no
# application-level auth scheme. ochakai reads the forwarded identity
# only to record provenance (design docs 0002/0003).
paths:
  /api/v1/knowledge:
    get:
      operationId: searchKnowledge
      summary: Search knowledge (trigram, plus hybrid vector search when configured)
      description: >
        Rejected entries are excluded unless the status filter explicitly
        includes rejected (so agents can check for already-rejected proposals).
        With sort=verified_at the endpoint lists entries by verification age
        (oldest first, never-verified last) instead of searching — the feed
        for golden query canary runs (docs/guides/golden-query-canary.md).
      parameters:
        - name: q
          in: query
          schema: { type: string }
        - name: sort
          in: query
          schema: { type: string, enum: [verified_at] }
        - name: type
          in: query
          schema: { type: array, items: { $ref: "#/components/schemas/Type" } }
          explode: true
        - name: status
          in: query
          schema: { type: array, items: { $ref: "#/components/schemas/Status" } }
          explode: true
        - name: tag
          in: query
          schema: { type: array, items: { type: string } }
          explode: true
        - name: limit
          in: query
          schema: { type: integer, default: 10, maximum: 50 }
      responses:
        "200":
          description: Scored hits, verified entries boosted.
          content:
            application/json:
              schema:
                type: object
                properties:
                  hits:
                    type: array
                    items: { $ref: "#/components/schemas/SearchHit" }
    post:
      operationId: createKnowledge
      summary: Create a knowledge entry
      description: >
        Entries default to draft. Anyone who can reach ochakai may set
        status=verified; verified_by records who did (provenance over
        authorization).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Knowledge" }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Knowledge" }
        "403": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /api/v1/context:
    get:
      operationId: getContext
      summary: The one-call read before answering a data question
      description: >
        Searches the knowledge base (verified entries rank higher), returns
        the full entries behind the top hits, and expands one hop through
        links so companion knowledge (the insight explaining a metric, the
        golden query answering the question) arrives in the same response.
        Rejected entries never appear. limit caps the primary entries
        (default 5, max 20); linked companions share a 2×limit total cap.
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: type
          in: query
          schema: { type: array, items: { $ref: "#/components/schemas/Type" } }
          explode: true
        - name: status
          in: query
          schema: { type: array, items: { $ref: "#/components/schemas/Status" } }
          explode: true
        - name: tag
          in: query
          schema: { type: array, items: { type: string } }
          explode: true
        - name: limit
          in: query
          schema: { type: integer, default: 5, maximum: 20 }
        - name: min_score
          in: query
          description: >
            Drop hits scoring below this before expansion, for callers that
            inject the pack automatically and prefer nothing over junk.
            Scores are search-mode dependent and uncalibrated (trigram
            similarity plus boosts vs RRF rank fusion) — calibrate against
            your own corpus before use.
          schema: { type: number, default: 0 }
      responses:
        "200":
          description: Ranked hits plus the full entries behind the top ones.
          content:
            application/json:
              schema:
                type: object
                properties:
                  hits:
                    type: array
                    items: { $ref: "#/components/schemas/SearchHit" }
                  entries:
                    type: array
                    items: { $ref: "#/components/schemas/Knowledge" }
        "400": { $ref: "#/components/responses/Error" }
  /api/v1/knowledge/{type}/{id}:
    parameters:
      - name: type
        in: path
        required: true
        schema: { $ref: "#/components/schemas/Type" }
      - name: id
        in: path
        required: true
        description: >
          Hierarchical IDs keep their slashes as real path separators
          (GET /api/v1/knowledge/query/sales/monthly-revenue).
        schema: { type: string }
    get:
      operationId: getKnowledge
      summary: Get one knowledge entry
      responses:
        "200":
          description: The entry.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Knowledge" }
        "404": { $ref: "#/components/responses/Error" }
    put:
      operationId: updateKnowledge
      summary: Update a knowledge entry (full replacement; revisions retained)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Knowledge" }
      responses:
        "200":
          description: Updated.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Knowledge" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      operationId: deleteKnowledge
      summary: Soft-delete a knowledge entry (history retained)
      responses:
        "204": { description: Deleted. }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/usage/{type}/{id}:
    parameters:
      - name: type
        in: path
        required: true
        schema: { $ref: "#/components/schemas/Type" }
      - name: id
        in: path
        required: true
        description: Hierarchical IDs keep their slashes as real path separators.
        schema: { type: string }
    get:
      operationId: getKnowledgeUsage
      summary: Usage totals for one knowledge entry
      description: >
        How often the entry was actually used: appeared in search results,
        fetched individually, or referenced by compile_sql. The measure of
        the write-back loop — evidence for promoting drafts, and a staleness
        signal for verified entries that stopped being used.
        (GET /api/v1/knowledge/{type}/{id}/usage remains as a legacy alias
        for single-segment IDs.)
      responses:
        "200":
          description: Usage totals.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Usage" }
        "404": { $ref: "#/components/responses/Error" }
  /api/v1/export:
    get:
      operationId: exportOKF
      summary: Export the knowledge base as an OKF bundle
      description: >
        Streams a tar.gz of the whole knowledge base in Open Knowledge
        Format (markdown + YAML frontmatter, one concept per entry, with
        index.md files). Your knowledge is yours.
      responses:
        "200":
          description: OKF bundle.
          content:
            application/gzip:
              schema: { type: string, format: binary }
  /api/v1/compile:
    post:
      operationId: compileSQL
      summary: Deterministically compile metrics into SQL
      description: >
        Compiles from the imported Ossie semantic model. ochakai never
        executes the SQL and never guesses: requests outside the supported
        subset (single fact + star joins) return 422 with a reason. Any
        verified golden queries about the metrics are returned alongside and
        should be preferred when they answer the question.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CompileRequest" }
      responses:
        "200":
          description: Compiled SQL.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompileResult" }
        "422": { $ref: "#/components/responses/Error" }
components:
  responses:
    Error:
      description: Error.
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string }
  schemas:
    Type:
      type: string
      description: >
        Any slug segment works as a type — the built-in five (metric, query,
        insight, term, table) are recommendations with server behavior
        attached, not a closed set (design doc 0005).
      examples: [metric, query, insight, term, table, runbook]
      pattern: "^[a-zA-Z0-9][a-zA-Z0-9._-]{0,127}$"
    Status:
      type: string
      description: >
        deprecated means "was correct, no longer recommended"; rejected
        means "was never accepted" — the record keeps agents from
        re-proposing the same knowledge.
      enum: [draft, verified, deprecated, rejected]
    Actor:
      type: object
      properties:
        kind: { type: string, enum: [human, agent] }
        name: { type: string }
    Link:
      type: object
      properties:
        rel: { type: string }
        target: { type: string, description: "<type>/<id>" }
    Knowledge:
      type: object
      required: [type, id, title]
      properties:
        type: { $ref: "#/components/schemas/Type" }
        id:
          type: string
          description: >
            Slug segments separated by "/" (hierarchical, mirroring OKF
            concept IDs); the final segment must not be "index".
          pattern: "^[a-zA-Z0-9][a-zA-Z0-9._-]{0,127}(/[a-zA-Z0-9][a-zA-Z0-9._-]{0,127})*$"
        title: { type: string }
        description: { type: string }
        tags: { type: array, items: { type: string } }
        status: { $ref: "#/components/schemas/Status" }
        status_note:
          type: string
          description: Free-form reason for the current status (why rejected/deprecated).
        created_by: { $ref: "#/components/schemas/Actor", readOnly: true }
        verified_by: { $ref: "#/components/schemas/Actor", readOnly: true }
        verified_at: { type: string, format: date-time, readOnly: true }
        rejected_by: { $ref: "#/components/schemas/Actor", readOnly: true }
        rejected_at: { type: string, format: date-time, readOnly: true }
        links: { type: array, items: { $ref: "#/components/schemas/Link" } }
        attrs:
          type: object
          description: Type-specific structured attributes (e.g. question/sql for query, model/expression for metric).
        body: { type: string, description: Markdown body. }
        created_at: { type: string, format: date-time, readOnly: true }
        updated_at: { type: string, format: date-time, readOnly: true }
    Usage:
      type: object
      properties:
        search_hits: { type: integer, description: Times the entry appeared in search results. }
        fetches: { type: integer, description: Times the entry was fetched individually. }
        compiles: { type: integer, description: Times compile_sql referenced the entry. }
        last_used_at: { type: string, format: date-time }
    SearchHit:
      allOf:
        - $ref: "#/components/schemas/Knowledge"
        - type: object
          properties:
            score: { type: number }
    CompileRequest:
      type: object
      required: [metrics]
      properties:
        model:
          type: string
          description: Semantic model name; optional when the first metric's knowledge entry links to its model.
        metrics: { type: array, items: { type: string }, minItems: 1 }
        dimensions:
          type: array
          items: { type: string, description: "dataset.field" }
        filters:
          type: array
          items:
            type: object
            required: [field, op, value]
            properties:
              field: { type: string, description: "dataset.field" }
              op: { type: string, enum: ["=", "!=", ">", ">=", "<", "<=", "in", "not_in"] }
              value: {}
        time_grain:
          type: object
          required: [field, grain]
          properties:
            field: { type: string, description: "dataset.field (time column)" }
            grain: { type: string, enum: [day, week, month, quarter, year] }
        dialect: { type: string, enum: [bigquery, ansi], default: bigquery }
        limit: { type: integer }
    CompileResult:
      type: object
      properties:
        sql: { type: string }
        dialect: { type: string }
        datasets_used: { type: array, items: { type: string } }
        notes: { type: array, items: { type: string } }
        verified_queries:
          type: array
          description: Verified golden queries about the requested metrics; prefer these when applicable.
          items: { $ref: "#/components/schemas/SearchHit" }